refactor: move run_review orchestration to review/runner.py
